Corrected C-spec Conversions
* Result indicators on file operations were not set correctly.

* Handling of DO was not outputting the correct end operations

* Fixes BrianGarland/vscode-rpgfree#91 fix which introduced a
bug that would cause the conversion of EVAL lines ending
with the operation code of  `+` or `-` to dropped.  This
was because the previous fix assumed _all_ lines ending with
a `+` or `-` were continuations of a literal.  The program now
checks for an open quote to confirm the continuation line.

In addition to the above fixes, the C.js source has been formatted
in a consistent manner; mostly conforming to the JS semistandard
formatting.  Notable exceptoins are the continued (but now
consistent use of template literals).