Additional handling of the conversion of DO/ENDDO
When converting a DO/ENDDO from fixed format to free format,
the loop is converted to a FOR/ENDFOR loop.  Previously, it was
assumed that the increment value of 1 was to be used.  This
change will attempt to insert the increment value found on the
matching ENDDO into the generated FOR statement.

Additionally, the fixed format DO statement supports the
ommission of `factor1`, 'factor2`, and/or `result`.  The
default value for `factor1` and `factor2` is 1.  If the
`result` field is omitted, RPG will auto-create a temporary
(internal) index variable.  In the free format, the index
variable is required.  The _Convert to Free Format_ action
will now auto-generate a (hopefully unique) variable name
to be used in the FOR/ENDFOR loop.  **The user is required
to define this variable.**

Also peformed some additional clean-up of the source code as
recommended by esLint using a semistandard configuration
with `curly` enabled.