Compile-time table and free form fixes
* Previously, no exception was made when converting all
source lines.  As a result, compile-time tables would be
parsed and converted.  This change will stop parsing
(converting) source lines once a compile-time table
is detected.

* When all lines to be converted were free form, an
error would be silently thrown which would prevent
the replacement of the converted source.  This error is
no longer being triggered, allowing the converted source
to be updated in the editor.
  * Fixes BrianGarland/vscode-rpgfree#53

* When conversion is run on the entire document,
the `**FREE` line was always being added as the first line.  This
would cause a new `**FREE` to be added each time the _Convert
to Free Format_ was run when no text was selected.  Changed to
only add the `**FREE` if the first line was not already the
`**FREE` marker.

* Removed an unecessary configuration setting that was erroneously
added.