Corrections to LOOKUP; add support for COMP, ...
The LOOKUP conversion assumed an equal condition only, ignoring
the HI and LO result indicators.

* LOOKUP will now be converted to the appropriate `%LookupXX` BIF.
* Results indicators will be correctly set.
* The LOOKUP will be converted to a table lookup (`%TlookupXX`)
if `factor2` is a table (variable name starts with TAB).  If
a `result` field was sepecified, it will be added to the
`%TlookupXX` BIF as appropriate.

Added initial support for conversion of COMP.

* Fixes BrianGarland/vscode-rpgfree#16

The math operations DIV and MULT did not support the
conversion of these operations when a `factor1` value
was not specified.  This change corrects these conversions.

The same indicator can be used for more than one result
condition (e.g., you can specify *in01 be used for HI, LO,
and/or EQ).  When converting a fixed format statement that
sets more than one result indicator, a new line of code was
added to set each result indicator.  This would cause and
indicator sepecified for more than one result condition to
reflect the last (HI, LO, EQ) condition only.

* The same result indicator can now be used in more than
one of the indicator positions and will properly refelct if
_any_ of the conditions being met.

Math functions support the use of result indicators to be
set on if the result was positive (+/HI), negative (-/LO)
and/or zero (Z/EQ).  Previously the result indicators on
math operations were being dropped upon conversion.  This
change adds support for result indicators on the following
math operations:

* `ADD`
* `DIV`
* `MULT
* `SUB`
* `XFOOT`
* `Z-ADD`
* `Z-SUB`

> **WARNING**
> The `MOVE`, `MOVEA`, and `MOVEL` support the result
> indicators `+` (positive), `-` (negative), or `ZB` (zero
> or blank).  At this time, these indicators are sitll
> being ignored.  Any use of result indicators for
> these operations will continue to drop references
> to the indicators on the converted source.