Significant Refactoring of C-Spec Conversion
> **Caution**
> This commit includes changes to nearly
> all of the C-spec op-code handlers.  Although
> limited unit testing has been done, more
> comprehensive testing is needed.  As such, this
> commit is being done to set a baseline and
> additional fixes are expected to be committed
> before this fork is ready to be merged.

This commit has multiple changes across the entire
project.  The primary focus was on refactoring the
C-spec conversion to ensure support for result indicators,
especially the ER result indicator.  In addition, the
conversion of each op-code was reviewed, updated, and
documented.

An overview of the changes:

* Updated the target environment to ECMAAScript 2022.

* Used ESLint on all source files.

* Fixed various conversion issues caused by cached
lines not being flushed (e.g., if the last line of
code was a `CALL`, the call would be dropped from
the source).

* Added support for conversion of `ALLOC`, `DEALLOC`,
`REALLOC`, `SHTDN`, and other op-codes.

* **Removed** the conversion of the `CAT` op-code
as we cannot guarantee the converted source will
function the same as the traditional syntax.

* Updated the README.md file with a table
that lists the level of support for each of the
traditional (fixed format) op-codes.